src/lambda/leakautomation/requirements.txt
PyGithub
python-dotenv
python-gitlab